requirements.txt
astroid==2.4.1
beautifulsoup4==4.8.2
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.2
pandas==1.0.3
pandas-datareader==0.8.1
pylint==2.5.2
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
